Stock-Market-Data/history_requirements.txt
requests==2.22.0
pandas==1.0.3
numpy==1.16.4
pyarrow==0.16.0
beautifulsoup4==4.9.0
protobuf==3.11.3
google-cloud-storage==1.28.0
